+++ static/i18n/locales/de/common.json
@@ -1,0 +1,1 @@
+{"'recent-workflows'":"'Recent Workflows'","'workflows-count_one'":"'{{count}} workflow'","'workflows-count_other'":"'{{count}} workflows'","'filtered-workflows-count'":"'Results {{filtered}} of $t(workflows-count, { \"count\": {{total}} })'"}
