+++ vite.config.ts
@@ -2,17 +2,9 @@ import path from 'path';
 import { defineConfig } from 'vite';
 import { sveltekit } from '@sveltejs/kit/vite';
 import { temporalServer } from './plugins/vite-plugin-temporal-server';
-import i18nPlugin from './plugins/vite-plugin-i18n';
 
 export default defineConfig({
-  plugins: [
-    sveltekit(),
-    temporalServer(),
-    i18nPlugin({
-      srcDir: path.resolve('./src/lib/i18n/locales'),
-      destDir: 'i18n/locales',
-    }),
-  ],
+  plugins: [sveltekit(), temporalServer()],
   optimizeDeps: {
     include: ['date-fns', 'date-fns-tz', 'websocket-as-promised'],
   },
